vectors-search-poc/requirements.txt
sentence-transformers
faiss-cpu
fastapi
uvicorn